scripts/requirements.txt
# 신규 가입 무료 망원경 지급 스크립트용
supabase>=2.0.0
python-dotenv>=1.0.0
